+++ pkg/linter/lib/src/rules/prefer_const_constructors_in_immutables.dart
@@ -105,7 +105,7 @@ class _Visitor extends SimpleAstVisitor<void> {
     }
 
     // Constructor with implicit `super()` call.
-    var unnamedSuperConstructor = clazz.supertype?.constructors2
+    var unnamedSuperConstructor = clazz.supertype?.constructors
         .firstWhereOrNull((e) => e.name3 == 'new');
     return unnamedSuperConstructor != null && unnamedSuperConstructor.isConst;
   }