+++ pkg/analyzer/CHANGELOG.md
@@ -71,6 +71,7 @@
 * Deprecate `ExtensionTypeFragment.representation2`, use `representation` instead.
 * Deprecate `DartType.element3`, use `element` instead.
 * Deprecate `LibraryElement.loadLibraryFunction2`, use `loadLibraryFunction` instead.
+* Deprecate `InterfaceType.constructors2`, use `constructors` instead.
 
 ## 7.4.1
 * Restore `InstanceElement.augmented` getter.